+++ cpp/src/io/text/data_chunk_source_factories.cpp
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+#include "io/text/device_data_chunks.hpp"
+
 #include <cudf/detail/nvtx/ranges.hpp>
 #include <cudf/io/text/data_chunk_source_factories.hpp>
 
@@ -27,30 +29,6 @@
 namespace cudf::io::text {
 
 namespace {
-
-class device_span_data_chunk : public device_data_chunk {
- public:
-  device_span_data_chunk(device_span<char const> data) : _data(data) {}
-
-  [[nodiscard]] char const* data() const override { return _data.data(); }
-  [[nodiscard]] std::size_t size() const override { return _data.size(); }
-  operator device_span<char const>() const override { return _data; }
-
- private:
-  device_span<char const> _data;
-};
-
-class device_uvector_data_chunk : public device_data_chunk {
- public:
-  device_uvector_data_chunk(rmm::device_uvector<char>&& data) : _data(std::move(data)) {}
-
-  [[nodiscard]] char const* data() const override { return _data.data(); }
-  [[nodiscard]] std::size_t size() const override { return _data.size(); }
-  operator device_span<char const>() const override { return _data; }
-
- private:
-  rmm::device_uvector<char> _data;
-};
 
 /**
  * @brief A reader which produces owning chunks of device memory which contain a copy of the data
